backend/pyproject.toml
[tool.poetry]
name = "fred"
version = "0.1-dev"
description = "Fred backend"
authors = ["Fred"]

[tool.poetry.dependencies]
python = "3.12.8"
fastapi = "0.115.6"
PyYAML = "6.0.1"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
croniter = "2.0.2"
pandas = "^2.2.2"
kubernetes = "30.1.0"
langchain = "^0.3.21"
langchain-openai = "^0.3.9"
langchain-core = "^0.3.47"
langchain-community = "^0.3.20"
langchain-mcp-adapters = "^0.0.11"
langgraph = "^0.2.34"
langfuse = "^2.52.0"
openai = "^1.54.0"
ipython = "^8.30.0"
pyjwt = "^2.10.1"
requests = "^2.32.3"
pydantic-settings = "^2.7.1"
langchain-ollama = "^0.3.0"
minio = "^7.2.15"
pytest-asyncio = "^0.26.0"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pip-audit = "^2.7.3"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry==1.7.1"]
build-backend = "poetry.core.masonry.api"
